backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
beautifulsoup4
requests
python-dotenv
google-genai
slowapi
